// src/services/generated/tainacanV2.schemas.ts
export type ListItemsParams = {
/**
 * Limits the result set to items with a specific title
 */
title?: string;
/**
 * Limit result to objects with specific id.
 */
id?: number;
/**
 * The context in which the request is made.
 */
context?: ListItemsContext;
/**
 * Limit results to those matching a string.
 */
search?: string;
/**
 * Whether to treat the search string as a single phrase.
 */
sentence?: boolean;
/**
 * Limit result set to objects assigned to specific authors by id.
 */
authorid?: number;
/**
 * Limit result set to objects assigned to specific authors by name
 */
authorname?: string;
/**
 * Limit result set to objects assigned one or more statuses.
 */
status?: ListItemsStatusItem[];
/**
 * Offset the result set by a specific number of objects.
 */
offset?: number;
/**
 * Order sort attribute ascending or descending.
 */
order?: 'asc' | 'desc' | 'ASC' | 'DESC' | unknown[];
/**
 * Sort objects by object attribute.
 */
orderby?: 'author' | 'date' | 'id' | 'include' | 'modified' | 'parent' | 'relevance' | 'slug' | 'include_slugs' | 'title' | 'meta_value' | 'meta_value_num' | 'menu_order' | 'rand' | 'post__in' | unknown[];
/**
 * Maximum number of objects to be returned in result set.
 */
perpage?: number;
/**
 * The results page to be returned.
 */
paged?: number;
/**
 * Ensure result set excludes specific IDs.
 */
exclude?: number[];
/**
 * Custom metadata key.
 */
metakey?: number | string;
/**
 * Custom metadata value
 */
metavalue?: string | unknown[];
/**
 * Custom metadata value
 */
metavaluenum?: number;
/**
 * Operator to test the metavalue
 */
metacompare?: ListItemsMetacompare;
/**
 * Limits result set to items that have specific custom metadata
 */
metaquery?: unknown[] | { [key: string]: unknown };
/**
 * Limits the result set to items that were created or modified in some specific date
 */
datequery?: unknown[] | { [key: string]: unknown };
/**
 * Show items associated with certain taxonomy.
 */
taxquery?: unknown[] | { [key: string]: unknown };
};

export type ListItemsContext = typeof ListItemsContext[keyof typeof ListItemsContext];


export const ListItemsContext = {
  view: 'view',
  edit: 'edit',
} as const;

export type ListItemsStatusItem = typeof ListItemsStatusItem[keyof typeof ListItemsStatusItem];


export const ListItemsStatusItem = {
  publish: 'publish',
  future: 'future',
  draft: 'draft',
  pending: 'pending',
  private: 'private',
  trash: 'trash',
  'auto-draft': 'auto-draft',
  inherit: 'inherit',
  'request-pending': 'request-pending',
  'request-confirmed': 'request-confirmed',
  'request-failed': 'request-failed',
  'request-completed': 'request-completed',
  any: 'any',
} as const;

export type ListItemsMetacompare = typeof ListItemsMetacompare[keyof typeof ListItemsMetacompare];


export const ListItemsMetacompare = {
  '=': '=',
  '!=': '!=',
  '>': '>',
  '>=': '>=',
  '<': '<',
  '<=': '<=',
  LIKE: 'LIKE',
  NOT_LIKE: 'NOT LIKE',
  IN: 'IN',
  NOT_IN: 'NOT IN',
  BETWEEN: 'BETWEEN',
  NOT_BETWEEN: 'NOT BETWEEN',
  NOT_EXISTS: 'NOT EXISTS',
  REGEXP: 'REGEXP',
  NOT_REGEXP: 'NOT REGEXP',
  RLIKE: 'RLIKE',
} as const;